Extract parseResult into a separate helper module
