+++ jest.config.js
@@ -2,4 +2,5 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   rootDir: 'src',
+  testPathIgnorePatterns: ['/node_modules/', '/__helpers__/'],
 };
